requirements.txt
django==1.3
fabric==1.2.2
psycopg2==2.4.1
python-memcached==1.47
django_compressor==1.1
sphinx==1.0.7
lxml==2.3.2
requests==0.8.3
python-dateutil==2.0
mock==0.7.2
geopy==0.94.1
django-tastypie==0.9.10
